requirements.txt
Flask
openai
tavily-python
requests
python-dotenv
pandas 